+++ src/candlewick/core/Components.h
@@ -28,4 +28,13 @@ struct MeshMaterialComponent {
   }
 };
 
+/// \brief Updates (adds or removes) the Opaque tag component for a given
+/// entity.
+/// \param reg The entity registry
+/// \param entity The entity
+/// \param mmc The mesh-material element to inspect for existence of any
+/// transparent subobjects.
+void updateTransparencyClassification(entt::registry &reg, entt::entity entity,
+                                      const MeshMaterialComponent &mmc);
+
 } // namespace candlewick